use new function readTableOrStop()
By doing so we can assure that we always use the same CSV settings (sep, dec, header,
...). In fact there is still a distinction between read.table with sep = ";" and dec = "."
and read.csv2, using sep = ";" and dec = ",", implicitly. This should not be the case.
Once all files are stored with the same CSV settings we can remove the argument "csv2"
and always use the read.table() function (TODO).
We also reuse the part of the error message that is always the same.