changeunit(): Lookup conversion factor in named vector
You can name the elements in a vector and access the
values by name instead of by index. If a name does
not exist, NA is returned by "[". In this case we
stop with an error message showing the units for which
conversion factors are defined (using stringList()
from kwb.utils).