remove arg "csv2" from readTableOrStop()
Instead, allow to pass additional arguments to read.table().
Do not call read.csv2() any more. As far as I can overlook
the only difference was the decimal character which now is set
to "," if read.csv2() was called before. Use the function
callWith() from the kwb.utils package to call a function with
a list of arguments, allowing to override general arguments
with specific ones (here: dec = "," overrides the default
dec = "."0)