rewrite summarise_loads() and use it in remaining cases
summarise_loads() now gets two main parameters:
- inputs is a list of named elements each of which is an data frame
  from which to copy values into the output data frames.
- columns is the vector of columns to be created in the output
  data frames.

Instead of the outer loop, use lapply(). It returns a list of which
we finally set the element names with structure(x, names = ...).

The most inner loop loops through the colum names (not indices).
The column name is split at the underscore character "_".
The first part is used to select the correct input data frame from
the list "inputs" by name and the second part (which must be one of
"mean", "5", "95") is converted to 1, 2 or three and then used as
row index "row.from" in the input data frame from where to copy the
value.