extract common code section to sumPaths()
Pass the input structures from which to sum up values in the list "inputs"
to sumPaths(). Within the inner loop of sumPaths(), use lapply() over all
inputs to select the appropriate columns. The result is "vectors", a list
of vectors of which a sum vector is calculated with Reduce("+", vectors).