improve getLoads()
Get "SUW_Names" out of the loops since they do not change within
the loops. Also outside the loops, filter the "volume" data frame
already for the given parameter. No need for which() when filtering
for the SUW name in "volume". Do this filtering already out of the
"run"-loop and only select the correct column (2 + run) within the
loop.