use lapply() in getLoads(), improve readability
Only now the unit change is actually applied because its result
is returned by the function called by lapply! Before, however,
changeunit() was called but the result was not assigned to any
variable! Set the removal rate to either 0 or the looked up rate,
in a one-line statement. Set the list element names at the end
with structure(x, names = ...)