remove arg "runs" from sumPaths()
The number of rows in each data frame of the output list is the same
as the number of rows in each input data frame. So there is no need
to give the number of runs as argument. We get the number of rows
from the number of rows in the first input (after checking if the
number of rows are equal in each input).